Renaming application folders
