+++ src/FluxServerFramework/assets/images/favicon.ico
