+++ src/SayHello/Presentation/home.html
@@ -1,4 +1,4 @@
-{% extends 'fluxFramework/viewTemplates/page.html' %}
+{% extends 'FluxServerFramework/viewTemplates/page.html' %}
 {% block content %}
     <div style="background-color: lightgray; text-align: center; min-height: calc(100vh - (100vh - 300px) / 2); padding-top: calc((100vh - 300px) / 2);">
         <img src="/assets/images/clean-code.png" /><h1>{{message}}</h1>
