+++ test/integration/test_TestMessageService.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 
 from src.SayHello.Domain.messageModel import messages
-from src.SayHello.Application.MessageService import MessageService
+from src.SayHello.ApplicationLogic.MessageService import MessageService
 
 class TestMessageService(TestCase):
